Porting tests to spock
